update terminal output
