+++ .snippets/code/tutorials/by-product/wormholescan/replace-signatures/replace-sigs-10.html
@@ -1,0 +1,6 @@
+<div id="termynal" data-termynal>
+	<span data-ty="input"><span class="file-path"></span>npx tsx test/checkVaaValidity.run.ts</span>
+	<span data-ty> </span>
+	<span data-ty>✅ VAA Valid: true</span>
+	<span data-ty="input"><span class="file-path"></span></span>
+</div>
