+++ .snippets/code/tutorials/by-product/wormholescan/replace-signatures/replace-sigs-11.html
@@ -1,0 +1,9 @@
+<div id="termynal" data-termynal>
+	<span data-ty="input"><span class="file-path"></span>npx tsx test/checkVaaValidity.run.ts</span>
+	<span data-ty> </span>
+	<span data-ty>❌ VAA Valid: false, Reason: VM signature invalid</span>
+	<span data-ty
+		>Transaction: 0x3ad91ec530187bb2ce3b394d587878cd1e9e037a97e51fbc34af89b2e0719367</span
+	>
+	<span data-ty="input"><span class="file-path"></span></span>
+</div>
