+++ .snippets/code/tutorials/by-product/wormholescan/replace-signatures/replace-sigs-14.ts
@@ -1,0 +1,10 @@
+import { fetchGuardianSet } from '../src/helpers/vaaHelper';
+
+const testFetchGuardianSet = async () => {
+  const [guardians, guardianSetIndex] = await fetchGuardianSet();
+
+  console.log('Current Guardian Set Index:', guardianSetIndex);
+  console.log('Guardian Addresses:', guardians);
+};
+
+testFetchGuardianSet();
