+++ .snippets/code/tutorials/by-product/wormholescan/replace-signatures/replace-sigs-15.html
@@ -1,0 +1,14 @@
+<div id="termynal" data-termynal>
+	<span data-ty="input"><span class="file-path"></span>npx tsx test/fetchGuardianSet.run.ts</span>
+	<span data-ty> </span>
+	<span data-ty>Fetching current guardian set</span>
+	<span data-ty>Current Guardian Set Index: 4</span>
+    <span data-ty>Guardian Addresses: [
+        '0x5893b5a76c3f739645648885bdccc06cd70a3cd3',
+        '0xff6cb952589bde862c25ef4392132fb9d4a42157',
+        '0x114de8460193bdf3a2fcf81f86a09765f4762fd1',
+        '0x107a0086b32d7a0977926a205131d8731d39cbeb',
+    </span>
+    <span data-ty>...]</span>
+	<span data-ty="input"><span class="file-path"></span></span>
+</div>
