+++ .snippets/code/tutorials/by-product/wormholescan/replace-signatures/replace-sigs-16.ts
