+++ .snippets/code/tutorials/by-product/wormholescan/replace-signatures/replace-sigs-17.html
@@ -1,0 +1,13 @@
+<div id="termynal" data-termynal>
+	<span data-ty="input"><span class="file-path"></span>npx tsx src/scripts/replaceSignatures.ts</span>
+	<span data-ty> </span>
+	<span data-ty>Processing TX: 0x3ad91ec530187bb2ce3b394d587878cd1e9e037a97e51fbc34af89b2e0719367</span>
+    <span data-ty>❌ VAA Valid: false, Reason: VM signature invalid</span>
+    <span data-ty>Fetching observations</span>
+    <span data-ty>Fetching current guardian set</span>
+    <span data-ty>Replacing Signatures...</span>
+    <span data-ty>Outdated Guardian Indexes: [ 0 ]</span>
+    <span data-ty>Sending updated VAA to RPC...</span>
+    <span data-ty>Updated VAA (hex): 0x01000000040d010019447b72d51e33923a3d6b28496ccd3722d5f1e33e2...</span>
+	<span data-ty="input"><span class="file-path"></span></span>
+</div>
