+++ .snippets/code/tutorials/by-product/wormholescan/replace-signatures/replace-sigs-6.html
@@ -1,0 +1,11 @@
+<div id="termynal" data-termynal>
+	<span data-ty="input"><span class="file-path"></span>npx tsx test/fetchVaaId.run.ts</span>
+	<span data-ty> </span>
+	<span data-ty
+		>Transaction: 0x3ad91ec530187bb2ce3b394d587878cd1e9e037a97e51fbc34af89b2e0719367</span
+	>
+	<span data-ty
+		>VAA ID: 2/0000000000000000000000003ee18b2214aff97000d974cf647e7c347e8fa585/164170</span
+	>
+	<span data-ty="input"><span class="file-path"></span></span>
+</div>
