+++ .snippets/code/tutorials/by-product/wormholescan/replace-signatures/replace-sigs-7.ts
@@ -1,11 +1,7 @@
-import {
-  fetchVaaId,
-  fetchVaa,
-  checkVaaValidity,
-} from '../src/helpers/vaaHelper';
+import { fetchVaaId, fetchVaa } from '../src/helpers/vaaHelper';
 import { TXS } from '../src/config/constants';
 
-const testCheckVaaValidity = async () => {
+const testFetchVaa = async () => {
   for (const tx of TXS) {
     const vaaIds = await fetchVaaId([tx]);
 
@@ -15,20 +11,15 @@ const testCheckVaaValidity = async () => {
     }
 
     for (const vaaId of vaaIds) {
-      const vaaData = await fetchVaa([vaaId]);
+      const vaaBytes = await fetchVaa([vaaId]);
 
-      if (vaaData.length === 0 || !vaaData[0].vaaBytes) {
-        console.log(`VAA not found for ID: ${vaaId}`);
-        continue;
-      }
-
-      const result = await checkVaaValidity(vaaData[0].vaaBytes);
       console.log(
-        `Transaction: ${tx}\nVAA ID: ${vaaId}\nVAA Validity:`,
-        result
+        `Transaction: ${tx}\nVAA ID: ${vaaId}\nVAA Bytes: ${
+          vaaBytes.length > 0 ? vaaBytes[0].vaaBytes : 'Not found'
+        }`
       );
     }
   }
 };
 
-testCheckVaaValidity();
+testFetchVaa();
