+++ .snippets/code/tutorials/by-product/wormholescan/replace-signatures/replace-sigs-8.html
@@ -1,0 +1,11 @@
+<div id="termynal" data-termynal>
+	<span data-ty="input"><span class="file-path"></span>npx tsx test/fetchVaa.run.ts</span>
+	<span data-ty> </span>
+	<span data-ty
+		>Transaction: 0x3ad91ec530187bb2ce3b394d587878cd1e9e037a97e51fbc34af89b2e0719367</span
+	>
+	<span data-ty
+		>VAA Bytes: AQAAAAMNANQSwD/HRPcKp7Yxypl1ON8dZeMBzgYJrd2KYz6l9Tq9K9fj72fYJgkMeMaB9h...</span
+	>
+	<span data-ty="input"><span class="file-path"></span></span>
+</div>
